\docType{data}
\name{gewaesserlinien}
\alias{gewaesserlinien}
\title{Rivers in the city of Osnabrueck}
\source{
  Frida: Free Vector-Geodata Osnabrueck
  <http://frida.intevation.org/download.html>
}
\description{
  A dataset containing the rivers of the german city of
  Osnabrueck.
}
\details{
  \itemize{ \item glName name of the river. }
}
\keyword{datasets}

