\docType{data}
\name{vg2500_bld}
\alias{vg2500_bld}
\title{Federal states of Germany}
\source{
  Bundesamt fuer Kartographie und Geodaesie

  <http://www.arcgis.com/home/item.html?id=ae25571c60d94ce5b7fcbf74e27c00e0>

  Copyright: Bundesamt fuer Kartographie und Geodaesie,
  Frankfurt am Main, 2011
}
\description{
  A dataset containing the Federal states of Germany.
}
\details{
  \itemize{ \item GEN name of the federal state.  \item
  SHAPE_LENG length of the border.  \item SHAPE_AREA area
  of the shape. }
}
\keyword{datasets}

